Refactored file structure, added scripts for compiling and running, improved Readme, removed use of Class.forName by adding driver to pom.xml dependencies.
