+++ scripts/compile.sh
@@ -1,0 +1,2 @@
+cd ..
+mvn clean package
