+++ scripts/run.sh
@@ -1,0 +1,2 @@
+cd ../target
+java -jar CofreDigital-1.0-SNAPSHOT-jar-with-dependencies.jar
