+++ src/main/java/CofreDigital/TOTP/TOTPTest.java
@@ -1,4 +1,4 @@
-package CofreDigital;
+package CofreDigital.TOTP;
 
 import java.util.Date;
