+++ src/main/java/CofreDigital/Users/User.java
@@ -3,7 +3,7 @@
   Ricardo Bastos Leta Vieira, 2110526
 */
 
-package CofreDigital;
+package CofreDigital.Users;
 
 /*import com.eatthepath:otp-java;
 import com.google.zxing.BarcodeFormat;
